Cleared out unnecessary typedefs that confused the VC++ Compiler
It seems when building in QtCreator using Desktop Qt 5.7.0
MSVC2015_64bit the typedefs for the binaryTree and list classes
are breaking the build. Removing the typedefs and uppercasing the list
class to List fixes the issues. Besides the successful build removing
these typedefs is stylistically more appropriate in c++ and
stylistically more consistent overall. Why lowercase the list class
if you will immediately typedef it to uppercase?